chore: make bruksvilkår e2e tests multi-mode
