+++ e2e/README.md
@@ -94,10 +94,10 @@ Run the tests from root using `yarn e2e:dev` or `yarn e2e`.
 
 ## [multi-mode/bruksvilkar.spec.ts](./multi-mode/bruksvilkar.spec.ts)
 
-- [FHIR: when bruksvilkår toggle is on, user should be automatically redirected to bruksvilkår page](./multi-mode/bruksvilkar.spec.ts#L7)
-- [FHIR: accepting the bruksvilkår should allow you to return to the patient](./multi-mode/bruksvilkar.spec.ts#L20)
-- [Standalone: when bruksvilkår toggle is on, user should be automatically redirected to bruksvilkår page](./multi-mode/bruksvilkar.spec.ts#L7)
-- [Standalone: accepting the bruksvilkår should allow you to return to the patient](./multi-mode/bruksvilkar.spec.ts#L20)
+- [FHIR: when bruksvilkår toggle is on, user should be automatically redirected to bruksvilkår page](./multi-mode/bruksvilkar.spec.ts#L6)
+- [FHIR: accepting the bruksvilkår should allow you to return to the patient](./multi-mode/bruksvilkar.spec.ts#L19)
+- [Standalone: when bruksvilkår toggle is on, user should be automatically redirected to bruksvilkår page](./multi-mode/bruksvilkar.spec.ts#L6)
+- [Standalone: accepting the bruksvilkår should allow you to return to the patient](./multi-mode/bruksvilkar.spec.ts#L19)
 
 ## [multi-mode/form-aareg.spec.ts](./multi-mode/form-aareg.spec.ts)
